refactor: dom 선택하는 로직 리팩토링
- dom 선택하는 로직을 gameView에서 모두 진행하도록 변경
- dom 조작은 오직 gameView에서만 담당하려고 설계했기 때문에 해당 리팩토링 진행
