+++ src/view/selector.js
